More realistic example code for how to use HQDM with a MagmaCoreDatabase.
